switch fast sync to new implementation
