+++ blockchain/pool.go
@@ -1,331 +1,94 @@
 package blockchain
 
 import (
-	"errors"
 	"fmt"
-	"math"
-	"sync"
-	"sync/atomic"
-	"time"
+	"sort"
 
 	cmn "github.com/tendermint/tendermint/libs/common"
-	flow "github.com/tendermint/tendermint/libs/flowrate"
 	"github.com/tendermint/tendermint/libs/log"
-
 	"github.com/tendermint/tendermint/p2p"
 	"github.com/tendermint/tendermint/types"
 )
 
-/*
-eg, L = latency = 0.1s
-	P = num peers = 10
-	FN = num full nodes
-	BS = 1kB block size
-	CB = 1 Mbit/s = 128 kB/s
-	CB/P = 12.8 kB
-	B/S = CB/P/BS = 12.8 blocks/s
-
-	12.8 * 0.1 = 1.28 blocks on conn
-*/
-
-const (
-	requestIntervalMS         = 2
-	maxTotalRequesters        = 600
-	maxPendingRequests        = maxTotalRequesters
-	maxPendingRequestsPerPeer = 20
-
-	// Minimum recv rate to ensure we're receiving blocks from a peer fast
-	// enough. If a peer is not sending us data at at least that rate, we
-	// consider them to have timedout and we disconnect.
-	//
-	// Assuming a DSL connection (not a good choice) 128 Kbps (upload) ~ 15 KB/s,
-	// sending data across atlantic ~ 7.5 KB/s.
-	minRecvRate = 7680
-
-	// Maximum difference between current and new block's height.
-	maxDiffBetweenCurrentAndReceivedBlockHeight = 100
-)
-
-var peerTimeout = 15 * time.Second // not const so we can override with tests
-
-/*
-	Peers self report their heights when we join the block pool.
-	Starting from our latest pool.height, we request blocks
-	in sequence from peers that reported higher heights than ours.
-	Every so often we ask peers what height they're on so we can keep going.
-
-	Requests are continuously made for blocks of higher heights until
-	the limit is reached. If most of the requests have no available peers, and we
-	are not at peer limits, we can probably switch to consensus reactor
-*/
-
-type BlockPool struct {
-	cmn.BaseService
-	startTime time.Time
-
-	mtx sync.Mutex
-	// block requests
-	requesters map[int64]*bpRequester
-	height     int64 // the lowest key in requesters.
-	// peers
-	peers         map[p2p.ID]*bpPeer
-	maxPeerHeight int64 // the biggest reported height
-
-	// atomic
-	numPending int32 // number of requests pending assignment or block response
-
-	requestsCh chan<- BlockRequest
-	errorsCh   chan<- peerError
+type blockData struct {
+	block *types.Block
+	peer  *bpPeer
 }
 
-// NewBlockPool returns a new BlockPool with the height equal to start. Block
-// requests and errors will be sent to requestsCh and errorsCh accordingly.
-func NewBlockPool(start int64, requestsCh chan<- BlockRequest, errorsCh chan<- peerError) *BlockPool {
-	bp := &BlockPool{
-		peers: make(map[p2p.ID]*bpPeer),
-
-		requesters: make(map[int64]*bpRequester),
-		height:     start,
-		numPending: 0,
-
-		requestsCh: requestsCh,
-		errorsCh:   errorsCh,
+func (bd *blockData) String() string {
+	if bd == nil {
+		return fmt.Sprintf("blockData nil")
 	}
-	bp.BaseService = *cmn.NewBaseService(nil, "BlockPool", bp)
-	return bp
+	if bd.block == nil {
+		if bd.peer == nil {
+			return fmt.Sprintf("block: nil peer: nil")
+		}
+		return fmt.Sprintf("block: nil peer: %v", bd.peer.id)
+	}
+	return fmt.Sprintf("block: %v peer: %v", bd.block.Height, bd.peer.id)
 }
 
-// OnStart implements cmn.Service by spawning requesters routine and recording
-// pool's start time.
-func (pool *BlockPool) OnStart() error {
-	go pool.makeRequestersRoutine()
-	pool.startTime = time.Now()
-	return nil
+type blockPool struct {
+	logger log.Logger
+	peers  map[p2p.ID]*bpPeer
+	blocks map[int64]p2p.ID
+
+	requests          map[int64]bool // list of blocks to be assigned peers for blockRequest
+	nextRequestHeight int64          // next request to be added to requests
+
+	height        int64 // processing height
+	maxPeerHeight int64 // maximum height of all peers
+	numPending    int32 // total numPending across peers
+	toBcR         bcRMessageInterface
 }
 
-// spawns requesters as needed
-func (pool *BlockPool) makeRequestersRoutine() {
-	for {
-		if !pool.IsRunning() {
-			break
-		}
-
-		_, numPending, lenRequesters := pool.GetStatus()
-		if numPending >= maxPendingRequests {
-			// sleep for a bit.
-			time.Sleep(requestIntervalMS * time.Millisecond)
-			// check for timed out peers
-			pool.removeTimedoutPeers()
-		} else if lenRequesters >= maxTotalRequesters {
-			// sleep for a bit.
-			time.Sleep(requestIntervalMS * time.Millisecond)
-			// check for timed out peers
-			pool.removeTimedoutPeers()
-		} else {
-			// request for more blocks.
-			pool.makeNextRequester()
-		}
+func newBlockPool(height int64, toBcR bcRMessageInterface) *blockPool {
+	return &blockPool{
+		peers:             make(map[p2p.ID]*bpPeer),
+		maxPeerHeight:     0,
+		blocks:            make(map[int64]p2p.ID),
+		requests:          make(map[int64]bool),
+		nextRequestHeight: height,
+		height:            height,
+		toBcR:             toBcR,
 	}
 }
 
-func (pool *BlockPool) removeTimedoutPeers() {
-	pool.mtx.Lock()
-	defer pool.mtx.Unlock()
-
-	for _, peer := range pool.peers {
-		if !peer.didTimeout && peer.numPending > 0 {
-			curRate := peer.recvMonitor.Status().CurRate
-			// curRate can be 0 on start
-			if curRate != 0 && curRate < minRecvRate {
-				err := errors.New("peer is not sending us data fast enough")
-				pool.sendError(err, peer.id)
-				pool.Logger.Error("SendTimeout", "peer", peer.id,
-					"reason", err,
-					"curRate", fmt.Sprintf("%d KB/s", curRate/1024),
-					"minRate", fmt.Sprintf("%d KB/s", minRecvRate/1024))
-				peer.didTimeout = true
-			}
-		}
-		if peer.didTimeout {
-			pool.removePeer(peer.id)
-		}
+func (pool *blockPool) String() string {
+	peerStr := fmt.Sprintf("Pool Peers:")
+	for _, p := range pool.peers {
+		peerStr += fmt.Sprintf("%v,", p)
 	}
+	return peerStr
+}
+
+func (pool *blockPool) setLogger(l log.Logger) {
+	pool.logger = l
 }
 
 // GetStatus returns pool's height, numPending requests and the number of
-// requesters.
-func (pool *BlockPool) GetStatus() (height int64, numPending int32, lenRequesters int) {
-	pool.mtx.Lock()
-	defer pool.mtx.Unlock()
-
-	return pool.height, atomic.LoadInt32(&pool.numPending), len(pool.requesters)
+// requests ready to be send in the future.
+func (pool *blockPool) getStatus() (height int64, numPending int32, maxPeerHeight int64) {
+	return pool.height, pool.numPending, pool.maxPeerHeight
 }
 
-// IsCaughtUp returns true if this node is caught up, false - otherwise.
-// TODO: relax conditions, prevent abuse.
-func (pool *BlockPool) IsCaughtUp() bool {
-	pool.mtx.Lock()
-	defer pool.mtx.Unlock()
-
-	// Need at least 1 peer to be considered caught up.
-	if len(pool.peers) == 0 {
-		pool.Logger.Debug("Blockpool has no peers")
-		return false
-	}
-
-	// Some conditions to determine if we're caught up.
-	// Ensures we've either received a block or waited some amount of time,
-	// and that we're synced to the highest known height.
-	// Note we use maxPeerHeight - 1 because to sync block H requires block H+1
-	// to verify the LastCommit.
-	receivedBlockOrTimedOut := pool.height > 0 || time.Since(pool.startTime) > 5*time.Second
-	ourChainIsLongestAmongPeers := pool.maxPeerHeight == 0 || pool.height >= (pool.maxPeerHeight-1)
-	isCaughtUp := receivedBlockOrTimedOut && ourChainIsLongestAmongPeers
-	return isCaughtUp
-}
-
-// We need to see the second block's Commit to validate the first block.
-// So we peek two blocks at a time.
-// The caller will verify the commit.
-func (pool *BlockPool) PeekTwoBlocks() (first *types.Block, second *types.Block) {
-	pool.mtx.Lock()
-	defer pool.mtx.Unlock()
-
-	if r := pool.requesters[pool.height]; r != nil {
-		first = r.getBlock()
-	}
-	if r := pool.requesters[pool.height+1]; r != nil {
-		second = r.getBlock()
-	}
-	return
-}
-
-// Pop the first block at pool.height
-// It must have been validated by 'second'.Commit from PeekTwoBlocks().
-func (pool *BlockPool) PopRequest() {
-	pool.mtx.Lock()
-	defer pool.mtx.Unlock()
-
-	if r := pool.requesters[pool.height]; r != nil {
-		/*  The block can disappear at any time, due to removePeer().
-		if r := pool.requesters[pool.height]; r == nil || r.block == nil {
-			PanicSanity("PopRequest() requires a valid block")
-		}
-		*/
-		r.Stop()
-		delete(pool.requesters, pool.height)
-		pool.height++
-	} else {
-		panic(fmt.Sprintf("Expected requester to pop, got nothing at height %v", pool.height))
-	}
-}
-
-// Invalidates the block at pool.height,
-// Remove the peer and redo request from others.
-// Returns the ID of the removed peer.
-func (pool *BlockPool) RedoRequest(height int64) p2p.ID {
-	pool.mtx.Lock()
-	defer pool.mtx.Unlock()
-
-	request := pool.requesters[height]
-	peerID := request.getPeerID()
-	if peerID != p2p.ID("") {
-		// RemovePeer will redo all requesters associated with this peer.
-		pool.removePeer(peerID)
-	}
-	return peerID
-}
-
-// TODO: ensure that blocks come in order for each peer.
-func (pool *BlockPool) AddBlock(peerID p2p.ID, block *types.Block, blockSize int) {
-	pool.mtx.Lock()
-	defer pool.mtx.Unlock()
-
-	requester := pool.requesters[block.Height]
-	if requester == nil {
-		pool.Logger.Info("peer sent us a block we didn't expect", "peer", peerID, "curHeight", pool.height, "blockHeight", block.Height)
-		diff := pool.height - block.Height
-		if diff < 0 {
-			diff *= -1
-		}
-		if diff > maxDiffBetweenCurrentAndReceivedBlockHeight {
-			pool.sendError(errors.New("peer sent us a block we didn't expect with a height too far ahead/behind"), peerID)
-		}
-		return
-	}
-
-	if requester.setBlock(block, peerID) {
-		atomic.AddInt32(&pool.numPending, -1)
-		peer := pool.peers[peerID]
-		if peer != nil {
-			peer.decrPending(blockSize)
-		}
-	} else {
-		pool.Logger.Info("invalid peer", "peer", peerID, "blockHeight", block.Height)
-		pool.sendError(errors.New("invalid peer"), peerID)
-	}
-}
-
-// MaxPeerHeight returns the highest reported height.
-func (pool *BlockPool) MaxPeerHeight() int64 {
-	pool.mtx.Lock()
-	defer pool.mtx.Unlock()
+func (pool blockPool) getMaxPeerHeight() int64 {
 	return pool.maxPeerHeight
 }
 
-// SetPeerHeight sets the peer's alleged blockchain height.
-func (pool *BlockPool) SetPeerHeight(peerID p2p.ID, height int64) {
-	pool.mtx.Lock()
-	defer pool.mtx.Unlock()
-
-	peer := pool.peers[peerID]
-	if peer != nil {
-		peer.height = height
-	} else {
-		peer = newBPPeer(pool, peerID, height)
-		peer.setLogger(pool.Logger.With("peer", peerID))
-		pool.peers[peerID] = peer
-	}
-
-	if height > pool.maxPeerHeight {
-		pool.maxPeerHeight = height
-	}
+func (pool *blockPool) reachedMaxHeight() bool {
+	return pool.maxPeerHeight == 0 || pool.height >= pool.maxPeerHeight
 }
 
-// RemovePeer removes the peer with peerID from the pool. If there's no peer
-// with peerID, function is a no-op.
-func (pool *BlockPool) RemovePeer(peerID p2p.ID) {
-	pool.mtx.Lock()
-	defer pool.mtx.Unlock()
-
-	pool.removePeer(peerID)
+func (pool *blockPool) rescheduleRequest(peerID p2p.ID, height int64) {
+	pool.logger.Debug("reschedule requests made to peer for height ", "peerID", peerID, "height", height)
+	pool.requests[height] = true
+	delete(pool.blocks, height)
+	delete(pool.peers[peerID].blocks, height)
 }
 
-func (pool *BlockPool) removePeer(peerID p2p.ID) {
-	for _, requester := range pool.requesters {
-		if requester.getPeerID() == peerID {
-			requester.redo(peerID)
-		}
-	}
-
-	peer, ok := pool.peers[peerID]
-	if ok {
-		if peer.timeout != nil {
-			peer.timeout.Stop()
-		}
-
-		delete(pool.peers, peerID)
-
-		// Find a new peer with the biggest height and update maxPeerHeight if the
-		// peer's height was the biggest.
-		if peer.height == pool.maxPeerHeight {
-			pool.updateMaxPeerHeight()
-		}
-	}
-}
-
-// If no peers are left, maxPeerHeight is set to 0.
-func (pool *BlockPool) updateMaxPeerHeight() {
+// Updates the pool's max height. If no peers are left maxPeerHeight is set to 0.
+func (pool *blockPool) updateMaxPeerHeight() {
 	var max int64
 	for _, peer := range pool.peers {
 		if peer.height > max {
@@ -335,296 +98,243 @@ func (pool *BlockPool) updateMaxPeerHeight() {
 	pool.maxPeerHeight = max
 }
 
-// Pick an available peer with at least the given minHeight.
-// If no peers are available, returns nil.
-func (pool *BlockPool) pickIncrAvailablePeer(minHeight int64) *bpPeer {
-	pool.mtx.Lock()
-	defer pool.mtx.Unlock()
+// Adds a new peer or updates an existing peer with a new height.
+// If the peer is too short it is removed.
+func (pool *blockPool) updatePeer(peerID p2p.ID, height int64) error {
+	pool.logger.Debug("updatePeer", "peerID", peerID, "height", height)
+	peer := pool.peers[peerID]
 
+	if height < pool.height {
+		pool.logger.Info("Peer height too small", "peer", peerID, "height", height, "fsm_height", pool.height)
+
+		// Don't add or update a peer that is not useful.
+		if peer != nil {
+			pool.logger.Info("remove short peer", "peer", peerID, "height", height, "fsm_height", pool.height)
+			pool.removePeer(peerID, errPeerTooShort)
+		}
+		return errPeerTooShort
+	}
+
+	if peer == nil {
+		// Add new peer.
+		peer = newBPPeer(peerID, height, pool.toBcR.sendPeerError)
+		peer.setLogger(pool.logger.With("peer", peerID))
+		pool.peers[peerID] = peer
+	} else {
+		// Update existing peer.
+		// Remove any requests made for heights in (height, peer.height].
+		for h, block := range pool.peers[peerID].blocks {
+			if h <= height {
+				continue
+			}
+			// Reschedule the requests for all blocks waiting for the peer, or received and not processed yet.
+			if block == nil {
+				// Since block was not yet received it is counted in numPending, decrement.
+				pool.numPending--
+				pool.peers[peerID].numPending--
+			}
+			pool.rescheduleRequest(peerID, h)
+		}
+		peer.height = height
+	}
+
+	pool.updateMaxPeerHeight()
+
+	return nil
+}
+
+// Stops the peer timer and deletes the peer. Recomputes the max peer height.
+func (pool *blockPool) deletePeer(peerID p2p.ID) {
+	if p, ok := pool.peers[peerID]; ok {
+		if p.timeout != nil {
+			p.timeout.Stop()
+		}
+		delete(pool.peers, peerID)
+
+		if p.height == pool.maxPeerHeight {
+			pool.updateMaxPeerHeight()
+		}
+	}
+}
+
+// Removes any blocks and requests associated with the peer and deletes the peer.
+// Also triggers new requests if blocks have been removed.
+func (pool *blockPool) removePeer(peerID p2p.ID, err error) {
+	pool.logger.Debug("removing peer", "peerID", peerID)
+
+	peer := pool.peers[peerID]
+	if peer == nil {
+		return
+	}
+	// Reschedule the requests for all blocks waiting for the peer, or received and not processed yet.
+	for h, block := range pool.peers[peerID].blocks {
+		if block == nil {
+			pool.numPending--
+		}
+		pool.rescheduleRequest(peerID, h)
+	}
+	pool.deletePeer(peerID)
+
+}
+
+// Called every time FSM advances its height.
+func (pool *blockPool) removeShortPeers() {
 	for _, peer := range pool.peers {
-		if peer.didTimeout {
-			pool.removePeer(peer.id)
+		if peer.height < pool.height {
+			pool.removePeer(peer.id, nil)
+		}
+	}
+}
+
+// Validates that the block comes from the peer it was expected from and stores it in the 'blocks' map.
+func (pool *blockPool) addBlock(peerID p2p.ID, block *types.Block, blockSize int) error {
+	if _, ok := pool.peers[peerID]; !ok {
+		pool.logger.Error("peer doesn't exist", "peer", peerID, "block_receieved", block.Height)
+		return errBadDataFromPeer
+	}
+	b, ok := pool.peers[peerID].blocks[block.Height]
+	if !ok {
+		pool.logger.Error("peer sent us a block we didn't expect", "peer", peerID, "blockHeight", block.Height)
+		if expPeerID, pok := pool.blocks[block.Height]; pok {
+			pool.logger.Error("expected this block from peer", "peer", expPeerID)
+		}
+		return errBadDataFromPeer
+	}
+	if b != nil {
+		pool.logger.Error("already have a block for height", "height", block.Height)
+		return errBadDataFromPeer
+	}
+
+	pool.peers[peerID].blocks[block.Height] = block
+	pool.blocks[block.Height] = peerID
+	pool.numPending--
+	pool.peers[peerID].decrPending(blockSize)
+	pool.logger.Debug("added new block", "height", block.Height, "from_peer", peerID, "total", len(pool.blocks))
+	return nil
+}
+
+func (pool *blockPool) getBlockAndPeerAtHeight(height int64) (bData *blockData, err error) {
+	peerID := pool.blocks[height]
+	peer := pool.peers[peerID]
+	if peer == nil {
+		return &blockData{}, errMissingBlocks
+	}
+
+	block, ok := peer.blocks[height]
+	if !ok || block == nil {
+		return &blockData{}, errMissingBlocks
+	}
+
+	return &blockData{peer: peer, block: block}, nil
+
+}
+
+func (pool *blockPool) getNextTwoBlocks() (first, second *blockData, err error) {
+	first, err = pool.getBlockAndPeerAtHeight(pool.height)
+	second, err2 := pool.getBlockAndPeerAtHeight(pool.height + 1)
+	if err == nil {
+		err = err2
+	}
+
+	if err == errMissingBlocks {
+		// We need both to sync the first block.
+		pool.logger.Error("missing blocks at height and/ or height+1", "height", pool.height)
+	}
+	return
+}
+
+// Remove peers that sent us the first two blocks, blocks will also be removed by removePeer().
+func (pool *blockPool) invalidateFirstTwoBlocks(err error) {
+	first, err1 := pool.getBlockAndPeerAtHeight(pool.height)
+	second, err2 := pool.getBlockAndPeerAtHeight(pool.height + 1)
+
+	if err1 == nil {
+		pool.removePeer(first.peer.id, err)
+	}
+	if err2 == nil {
+		pool.removePeer(second.peer.id, err)
+	}
+}
+
+func (pool *blockPool) processedCurrentHeightBlock() {
+	peerID, peerOk := pool.blocks[pool.height]
+	if peerOk {
+		delete(pool.peers[peerID].blocks, pool.height)
+	}
+	delete(pool.blocks, pool.height)
+	pool.logger.Debug("processed and removed block at height", "height", pool.height)
+	pool.height++
+	pool.removeShortPeers()
+}
+
+func (pool *blockPool) removeBadPeers() {
+	pool.removeShortPeers()
+	for _, peer := range pool.peers {
+		if err := peer.isGood(); err != nil {
+			pool.removePeer(peer.id, err)
+			if err == errSlowPeer {
+				peer.errFunc(errSlowPeer, peer.id)
+			}
+		}
+	}
+}
+
+func (pool *blockPool) makeRequestBatch(maxNumPendingRequests int32) []int {
+	pool.removeBadPeers()
+	// If running low on planned requests, make more.
+	numNeeded := int32(cmn.MinInt(int(maxNumPendingRequests), len(pool.peers)*int(maxRequestsPerPeer))) - pool.numPending
+	for int32(len(pool.requests)) < numNeeded {
+		if pool.nextRequestHeight > pool.maxPeerHeight {
+			break
+		}
+		pool.requests[pool.nextRequestHeight] = true
+		pool.nextRequestHeight++
+	}
+
+	heights := make([]int, 0, len(pool.requests))
+	for k := range pool.requests {
+		heights = append(heights, int(k))
+	}
+	sort.Ints(heights)
+	return heights
+}
+
+func (pool *blockPool) makeNextRequests(maxNumPendingRequests int32) {
+	heights := pool.makeRequestBatch(maxNumPendingRequests)
+	pool.logger.Debug("makeNextRequests will make following requests", "number", len(heights), "heights", heights)
+
+	for _, height := range heights {
+		h := int64(height)
+		if err := pool.sendRequest(h); err != nil {
+			// Errors from sendRequest() are handled by this function
+			return
+		}
+		delete(pool.requests, h)
+	}
+}
+
+func (pool *blockPool) sendRequest(height int64) error {
+	for _, peer := range pool.peers {
+		if peer.numPending >= int32(maxRequestsPerPeer) {
 			continue
 		}
-		if peer.numPending >= maxPendingRequestsPerPeer {
+		if peer.height < height {
 			continue
 		}
-		if peer.height < minHeight {
-			continue
+		pool.logger.Debug("assign request to peer", "peer", peer.id, "height", height)
+		if err := pool.toBcR.sendBlockRequest(peer.id, height); err == errNilPeerForBlockRequest {
+			pool.removePeer(peer.id, err)
+			pool.toBcR.sendPeerError(err, peer.id)
 		}
+
+		pool.blocks[height] = peer.id
+		pool.numPending++
+
+		peer.blocks[height] = nil
 		peer.incrPending()
-		return peer
+
+		return nil
 	}
-	return nil
-}
-
-func (pool *BlockPool) makeNextRequester() {
-	pool.mtx.Lock()
-	defer pool.mtx.Unlock()
-
-	nextHeight := pool.height + pool.requestersLen()
-	if nextHeight > pool.maxPeerHeight {
-		return
-	}
-
-	request := newBPRequester(pool, nextHeight)
-
-	pool.requesters[nextHeight] = request
-	atomic.AddInt32(&pool.numPending, 1)
-
-	err := request.Start()
-	if err != nil {
-		request.Logger.Error("Error starting request", "err", err)
-	}
-}
-
-func (pool *BlockPool) requestersLen() int64 {
-	return int64(len(pool.requesters))
-}
-
-func (pool *BlockPool) sendRequest(height int64, peerID p2p.ID) {
-	if !pool.IsRunning() {
-		return
-	}
-	pool.requestsCh <- BlockRequest{height, peerID}
-}
-
-func (pool *BlockPool) sendError(err error, peerID p2p.ID) {
-	if !pool.IsRunning() {
-		return
-	}
-	pool.errorsCh <- peerError{err, peerID}
-}
-
-// for debugging purposes
-//nolint:unused
-func (pool *BlockPool) debug() string {
-	pool.mtx.Lock()
-	defer pool.mtx.Unlock()
-
-	str := ""
-	nextHeight := pool.height + pool.requestersLen()
-	for h := pool.height; h < nextHeight; h++ {
-		if pool.requesters[h] == nil {
-			str += fmt.Sprintf("H(%v):X ", h)
-		} else {
-			str += fmt.Sprintf("H(%v):", h)
-			str += fmt.Sprintf("B?(%v) ", pool.requesters[h].block != nil)
-		}
-	}
-	return str
-}
-
-//-------------------------------------
-
-type bpPeer struct {
-	pool        *BlockPool
-	id          p2p.ID
-	recvMonitor *flow.Monitor
-
-	height     int64
-	numPending int32
-	timeout    *time.Timer
-	didTimeout bool
-
-	logger log.Logger
-}
-
-func newBPPeer(pool *BlockPool, peerID p2p.ID, height int64) *bpPeer {
-	peer := &bpPeer{
-		pool:       pool,
-		id:         peerID,
-		height:     height,
-		numPending: 0,
-		logger:     log.NewNopLogger(),
-	}
-	return peer
-}
-
-func (peer *bpPeer) setLogger(l log.Logger) {
-	peer.logger = l
-}
-
-func (peer *bpPeer) resetMonitor() {
-	peer.recvMonitor = flow.New(time.Second, time.Second*40)
-	initialValue := float64(minRecvRate) * math.E
-	peer.recvMonitor.SetREMA(initialValue)
-}
-
-func (peer *bpPeer) resetTimeout() {
-	if peer.timeout == nil {
-		peer.timeout = time.AfterFunc(peerTimeout, peer.onTimeout)
-	} else {
-		peer.timeout.Reset(peerTimeout)
-	}
-}
-
-func (peer *bpPeer) incrPending() {
-	if peer.numPending == 0 {
-		peer.resetMonitor()
-		peer.resetTimeout()
-	}
-	peer.numPending++
-}
-
-func (peer *bpPeer) decrPending(recvSize int) {
-	peer.numPending--
-	if peer.numPending == 0 {
-		peer.timeout.Stop()
-	} else {
-		peer.recvMonitor.Update(recvSize)
-		peer.resetTimeout()
-	}
-}
-
-func (peer *bpPeer) onTimeout() {
-	peer.pool.mtx.Lock()
-	defer peer.pool.mtx.Unlock()
-
-	err := errors.New("peer did not send us anything")
-	peer.pool.sendError(err, peer.id)
-	peer.logger.Error("SendTimeout", "reason", err, "timeout", peerTimeout)
-	peer.didTimeout = true
-}
-
-//-------------------------------------
-
-type bpRequester struct {
-	cmn.BaseService
-	pool       *BlockPool
-	height     int64
-	gotBlockCh chan struct{}
-	redoCh     chan p2p.ID //redo may send multitime, add peerId to identify repeat
-
-	mtx    sync.Mutex
-	peerID p2p.ID
-	block  *types.Block
-}
-
-func newBPRequester(pool *BlockPool, height int64) *bpRequester {
-	bpr := &bpRequester{
-		pool:       pool,
-		height:     height,
-		gotBlockCh: make(chan struct{}, 1),
-		redoCh:     make(chan p2p.ID, 1),
-
-		peerID: "",
-		block:  nil,
-	}
-	bpr.BaseService = *cmn.NewBaseService(nil, "bpRequester", bpr)
-	return bpr
-}
-
-func (bpr *bpRequester) OnStart() error {
-	go bpr.requestRoutine()
-	return nil
-}
-
-// Returns true if the peer matches and block doesn't already exist.
-func (bpr *bpRequester) setBlock(block *types.Block, peerID p2p.ID) bool {
-	bpr.mtx.Lock()
-	if bpr.block != nil || bpr.peerID != peerID {
-		bpr.mtx.Unlock()
-		return false
-	}
-	bpr.block = block
-	bpr.mtx.Unlock()
-
-	select {
-	case bpr.gotBlockCh <- struct{}{}:
-	default:
-	}
-	return true
-}
-
-func (bpr *bpRequester) getBlock() *types.Block {
-	bpr.mtx.Lock()
-	defer bpr.mtx.Unlock()
-	return bpr.block
-}
-
-func (bpr *bpRequester) getPeerID() p2p.ID {
-	bpr.mtx.Lock()
-	defer bpr.mtx.Unlock()
-	return bpr.peerID
-}
-
-// This is called from the requestRoutine, upon redo().
-func (bpr *bpRequester) reset() {
-	bpr.mtx.Lock()
-	defer bpr.mtx.Unlock()
-
-	if bpr.block != nil {
-		atomic.AddInt32(&bpr.pool.numPending, 1)
-	}
-
-	bpr.peerID = ""
-	bpr.block = nil
-}
-
-// Tells bpRequester to pick another peer and try again.
-// NOTE: Nonblocking, and does nothing if another redo
-// was already requested.
-func (bpr *bpRequester) redo(peerId p2p.ID) {
-	select {
-	case bpr.redoCh <- peerId:
-	default:
-	}
-}
-
-// Responsible for making more requests as necessary
-// Returns only when a block is found (e.g. AddBlock() is called)
-func (bpr *bpRequester) requestRoutine() {
-OUTER_LOOP:
-	for {
-		// Pick a peer to send request to.
-		var peer *bpPeer
-	PICK_PEER_LOOP:
-		for {
-			if !bpr.IsRunning() || !bpr.pool.IsRunning() {
-				return
-			}
-			peer = bpr.pool.pickIncrAvailablePeer(bpr.height)
-			if peer == nil {
-				//log.Info("No peers available", "height", height)
-				time.Sleep(requestIntervalMS * time.Millisecond)
-				continue PICK_PEER_LOOP
-			}
-			break PICK_PEER_LOOP
-		}
-		bpr.mtx.Lock()
-		bpr.peerID = peer.id
-		bpr.mtx.Unlock()
-
-		// Send request and wait.
-		bpr.pool.sendRequest(bpr.height, peer.id)
-	WAIT_LOOP:
-		for {
-			select {
-			case <-bpr.pool.Quit():
-				bpr.Stop()
-				return
-			case <-bpr.Quit():
-				return
-			case peerID := <-bpr.redoCh:
-				if peerID == bpr.peerID {
-					bpr.reset()
-					continue OUTER_LOOP
-				} else {
-					continue WAIT_LOOP
-				}
-			case <-bpr.gotBlockCh:
-				// We got a block!
-				// Continue the for-loop and wait til Quit.
-				continue WAIT_LOOP
-			}
-		}
-	}
-}
-
-//-------------------------------------
-
-type BlockRequest struct {
-	Height int64
-	PeerID p2p.ID
+	pool.logger.Error("could not find peer to send request for block at height", "height", height)
+	return errNoPeerFoundForHeight
 }
