+++ blockchain_old/store_test.go
@@ -1,4 +1,4 @@
-package blockchain_new
+package blockchain_old
 
 import (
 	"bytes"
